requirements.txt
numpy>=1.24.0
scipy>=1.10.0
cvxpy>=1.4.0
pandas>=2.0.0
matplotlib>=3.7.0
requests>=2.31.0
pyyaml>=6.0
python-telegram-bot>=20.0
